* commit 'd3560aac09d2fd88b6e695ec1a51b262b0922c07':
  CPS-1675: address PR feedback.
  CPS-1675: removed redundant code.
  CPS-1675: add a Kong-aware Pydap data store for xarray. A few other bug fixes crept in too.

+++ as_models/web_api.py
@@ -355,6 +355,9 @@ def _fail_with_exception(message, dev_message, data):
 
 
 def terminate(timeout=0.0):
+    if _process is None:
+        return  # Can't terminate model - it never started.
+
     _logger.debug('Waiting %.2f seconds for model to terminate.', timeout)
 
     _process.terminate()
